requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==0.61.1
djangorestframework==3.15.2
idna==3.10
oauthlib==3.2.2
psycopg==3.2.3
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
urllib3==2.3.0
